% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_final_jeapordy.R
\name{mod_final_jeapordy_ui}
\alias{mod_final_jeapordy_ui}
\alias{mod_final_jeapordy_server}
\title{mod_final_jeapordy_ui and mod_final_jeapordy_server}
\usage{
mod_final_jeapordy_ui(id)

mod_final_jeapordy_server(input, output, session, game_info, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
