% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ind_box.R
\name{mod_ind_box_ui}
\alias{mod_ind_box_ui}
\alias{mod_ind_box_server}
\title{mod_ind_box_ui and mod_ind_box_server}
\usage{
mod_ind_box_ui(id)

mod_ind_box_server(input, output, session, game_info, selected_row,
  selected_col, selected_round, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
